backend/requirements.txt
Django==4.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
requests==2.31.0
psycopg2==2.9.6
redis==4.6.0
celery==5.3.6